chore: renamed modules exceptions->errors; bedrock_models->deployments
