+++ aidial_adapter_bedrock/llm/model/stability.py
@@ -15,7 +15,7 @@ from aidial_adapter_bedrock.llm.chat_model import (
     TextCompletionPrompt,
 )
 from aidial_adapter_bedrock.llm.consumer import Attachment, Consumer
-from aidial_adapter_bedrock.llm.exceptions import ValidationError
+from aidial_adapter_bedrock.llm.errors import ValidationError
 from aidial_adapter_bedrock.llm.message import BaseMessage
 from aidial_adapter_bedrock.llm.tools.default_emulator import (
     default_tools_emulator,
